fix user input error during game, old game file deleted
