+++ lib/game.rb
@@ -1,69 +1,90 @@
-# class Game
-#
-#   def initialize
-#   end
-#
-#   def start
-#     puts "     Welcome to BATTLESHIP!!!! \n Enter p to play. Enter q to quit."
-#     answer = gets.chomp
-#
-#
-#
-#       if answer.downcase == "p"
-#           puts "I have laid out my ships on the grid. \nYou now need to lay out your two ships.\nThe Cruiser is three units long and the Submarine is two units long."
-#           @player_board = Board.new
-#           @computer_board = Board.new
-#           player_cruiser = Ship.new("Cruiser", 3)
-#           player_submarine = Ship.new("Submarine", 2)
-#           computer_cruiser = Ship.new("Cruiser", 3)
-#           computer_submarine = Ship.new("Submarine", 2)
-#           computer_choices = ComputerChoices.new
-#
-#
-#           @computer_board.place(computer_cruiser, computer_choices.randomly_generated_sub_array)
-#           @computer_board.place(computer_submarine, computer_choices.randomly_generated_cruiser_array)
-#
-#           puts @player_board.render
-#           puts "Enter the squares for the Cruiser in order (3 spaces) You can only place your ship vertically or horizontally:"
-#           user_choice_for_cruiser = gets.chomp.split(" ")
-#           unless @player_board.valid_placement?(player_cruiser, user_choice_for_cruiser)
-#             #not reaching this, yet!
-#             puts "That isn't a valid placement."
-#           else
-#             @player_board.place(player_cruiser, user_choice_for_cruiser)
-#             puts @player_board.render(true)
-#             puts "Place your 2nd ship called the Submarine (2 spaces) You cannot overlap ships, so don't try!"
-#             user_choice_for_submarine = gets.chomp.split(" ")
-#             unless @player_board.valid_placement?(player_submarine, user_choice_for_submarine)
-#               puts "That isn't a valid placement!"
-#             else
-#               @player_board.place(player_submarine, user_choice_for_submarine)
-#               puts @player_board.render(true)
-#             end
-#
-#           end
-#       elsif answer == "q"
-#         puts "Please remove yourself from the battlefield..."
-#       else
-#         puts "Please try again using either p to play or q to quit."
-#         answer = gets.chomp
-#       end
-#
-#
-#       puts "Now that you have placed your ships, it's time to strike."
-#       puts
-#       until player_cruiser.sunk == true && player_submarine.sunk == true || computer_cruiser.sunk == true && computer_submarine.sunk == true
-#         puts "Which coordinate do you think your opponent ship is on?"
-#         first_player_guess = gets.chomp
-#         @computer_board.cells[first_player_guess].fire_upon
-#         puts @computer_board.render
-#       end
-#   end
-#
-#
-#
-# end
-#
-# #I think this should be added to the unless statement for validating board placement but valid_coordinate
-# #can only be called on a board object & i am calling it on an array.
-# #&& user_choice_for_cruiser.all? {|coordinate| valid_coordinate?(coordinate)}
+class Game
+    attr_reader :player_board, :computer_board, :player_cruiser, :player_submarine,
+                :computer_cruiser, :computer_submarine, :computer_choices, :player_input,
+                :computer_guess_array
+
+  def initialize
+    @player_board = Board.new
+    @computer_board = Board.new
+    @player_cruiser = Ship.new("Cruiser", 3)
+    @player_submarine = Ship.new("Submarine", 2)
+    @computer_cruiser = Ship.new("Cruiser", 3)
+    @computer_submarine = Ship.new("Submarine", 2)
+    @player_input = nil
+    @computer_choices = nil
+    @computer_guess_array = ["A1", "A2", "A3", "A4", "B1", "B2", "B3", "B4", "C1", "C2", "C3", "C4", "D1", "D2", "D3", "D4"]
+  end
+
+  def run
+    until @player_input == "q"
+      puts "Welcome to BATTLESHIP!!!! \n Enter p to play. Enter q to quit."
+      @player_input = gets.chomp
+      if @player_input == "p"
+        run_computer_choices
+        play_game
+        return
+      end
+    end
+  end
+
+  def run_computer_choices
+      @computer_choices = ComputerChoices.new
+      until computer_choices.valid_placement_computer?
+        @computer_choices = ComputerChoices.new
+    end
+  end
+
+
+
+  def play_game
+    @computer_board.place(@computer_cruiser, @computer_choices.randomly_generated_cruiser_array)
+    @computer_board.place(@computer_submarine, @computer_choices.randomly_generated_sub_array)
+    puts "I have laid out my ships on the grid. \nYou now need to lay out your two ships.\nThe Cruiser is 3 units long and the Submarine is 2 units long."
+    puts @player_board.render(true)
+    puts "Enter the squares for the #{@player_cruiser.name} in order (#{@player_cruiser.length} spaces). You can only place your ship vertically or horizontally:"
+
+    user_placement_choice = gets.chomp.split(" ")
+    until @player_board.valid_placement?(@player_cruiser, user_placement_choice)
+      puts "That isn't a valid placement."
+      puts "Enter the squares for the #{@player_cruiser.name} in order (#{@player_cruiser.length} spaces). You can only place your ship vertically or horizontally:"
+      user_placement_choice = gets.chomp.split(" ")
+    end
+    @player_board.place(@player_cruiser, user_placement_choice)
+    puts @player_board.render(true)
+
+    puts "Enter the squares for the #{@player_submarine.name} in order (#{@player_submarine.length} spaces). You can only place your ship vertically or horizontally:"
+    user_placement_choice = gets.chomp.split(" ")
+
+    until @player_board.valid_placement?(@player_submarine, user_placement_choice)
+      puts "That isn't a valid placement."
+      puts "Enter the squares for the #{@player_submarine.name} in order (#{@player_submarine.length} spaces). You can only place your ship vertically or horizontally:"
+      user_placement_choice = gets.chomp.split(" ")
+    end
+    @player_board.place(@player_submarine, user_placement_choice)
+    puts @player_board.render(true)
+    puts "Now that you have placed your ships, it's time to strike."
+
+    until @player_cruiser.sunk == true && @player_submarine.sunk == true || @computer_cruiser.sunk == true && @computer_submarine.sunk == true
+      puts "Which coordinate do you think your opponent ship is on?"
+      first_player_guess = gets.chomp
+      if @computer_board.valid_coordinate?(first_player_guess)
+        @computer_board.cells[first_player_guess].fire_upon
+        puts "=============COMPUTER BOARD============="
+        puts @computer_board.render
+        puts
+        next_cell = @player_board.cells[@computer_guess_array.shuffle.pop]
+        next_cell.fire_upon
+        puts "I've taken my shot as well"
+        puts "==============PLAYER BOARD=============="
+        puts @player_board.render(true)
+      elsif
+        puts "Please enter a VALID coordinate"
+      end
+    end
+
+    puts "Game over!"
+
+
+  end
+
+end
