+++ lib/runner.rb
@@ -1,7 +1,7 @@
 require './lib/ship.rb'
 require './lib/cell.rb'
 require './lib/board.rb'
-require './lib/game2.rb'
+require './lib/game.rb'
 require './lib/computer_choices.rb'
